# Conflicts:
#	service/server/src/main/java/com/emc/pravega/service/server/containers/StreamSegmentContainer.java
#	service/server/src/main/java/com/emc/pravega/service/server/containers/StreamSegmentMapper.java
#	service/server/src/main/java/com/emc/pravega/service/server/store/StreamSegmentService.java
#	service/server/src/test/java/com/emc/pravega/service/server/containers/StreamSegmentContainerTests.java
#	service/server/src/test/java/com/emc/pravega/service/server/containers/StreamSegmentMapperTests.java
#	service/storage/src/main/java/com/emc/pravega/service/storage/mocks/InMemoryStorage.java

+++ service/server/src/main/java/com/emc/pravega/service/server/containers/StreamSegmentContainer.java
@@ -144,7 +144,7 @@ class StreamSegmentContainer extends AbstractService implements SegmentContainer
 
     @Override
     protected void doStart() {
-        int traceId = LoggerHelpers.traceEnter(log, traceObjectId, "doStart");
+        long traceId = LoggerHelpers.traceEnter(log, traceObjectId, "doStart");
         this.durableLog.startAsync();
         this.executor.execute(() -> {
             this.durableLog.awaitRunning();
@@ -161,7 +161,7 @@ class StreamSegmentContainer extends AbstractService implements SegmentContainer
 
     @Override
     protected void doStop() {
-        int traceId = LoggerHelpers.traceEnter(log, traceObjectId, "doStop");
+        long traceId = LoggerHelpers.traceEnter(log, traceObjectId, "doStop");
         this.writer.stopAsync();
         this.durableLog.stopAsync();
         this.executor.execute(() -> {
